# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mait AI Platform API
  version: 1.0.0
  description: Artificial Insemination field operations platform for Shwetdhara Milk
    Producer Company. Serves the Mait mobile app, the admin web portal and the Indent
    Easy integration.
paths:
...
  /api/v1/members/:
    get:
      operationId: members_list
      description: |-
        Member search (SRS §9.3).

        Backed by the ``member_mpp_name_idx`` composite index — the table holds 105k+ rows, so an
        unindexed search here would be the slowest thing a Mait does all day.
      summary: Search members
      parameters:
      - in: query
        name: activation_status
        schema:
          type: string
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: mobile_no
        schema:
          type: string
      - in: query
        name: mpp__mpp_code
        schema:
          type: string
        description: Restrict to one MPP. The mobile app always sends this.
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - in: query
        name: search
        schema:
          type: string
        description: Matches member name, member code or mobile number.
      tags:
      - master-data
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMemberListList'
          description: ''
...
  /api/v1/mpp/:
    get:
      operationId: mpp_list
      description: MPP directory (SRS §9.3).
      parameters:
      - in: query
        name: district_code
        schema:
          type: string
      - in: query
        name: is_active
        schema:
          type: boolean
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: mait
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: tehsil_code
        schema:
          type: string
      tags:
      - master-data
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMPPListList'
          description: ''
...
components:
  schemas:
...
    MPPList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        mpp_code:
          type: string
          readOnly: true
        mpp_name:
          type: string
          readOnly: true
        district_code:
          type: string
          readOnly: true
        tehsil_code:
          type: string
          readOnly: true
        village_code:
          type: string
          readOnly: true
        mobile_no:
          type: string
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
        mait:
          type: integer
          readOnly: true
          nullable: true
          description: Assigned Sahayak/Mait. Admin can override the SAP default (SRS
            §6.2.2).
        mait_name:
          type: string
          readOnly: true
      required:
      - district_code
      - id
      - is_active
      - mait
      - mait_name
      - mobile_no
      - mpp_code
      - mpp_name
      - tehsil_code
      - village_code
...
    MemberList:
      type: object
      description: |-
        Trimmed for the mobile app's member picker (SRS §6.3 step 2).

        A Mait standing at an MPP needs a name, a code and a number to confirm identity — not 54
        columns over a rural connection.
      properties:
        id:
          type: integer
          readOnly: true
        member_code:
          type: string
          readOnly: true
        member_name:
          type: string
          readOnly: true
        father_husband_name:
          type: string
          readOnly: true
        mobile_no:
          type: string
          readOnly: true
          description: Payment authorisation OTPs are sent here (SRS §6.5).
        mpp_code:
          type: string
          readOnly: true
        activation_status:
          type: string
          readOnly: true
      required:
      - activation_status
      - father_husband_name
      - id
      - member_code
      - member_name
      - mobile_no
      - mpp_code
...
    PaginatedMPPListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/MPPList'
    PaginatedMemberListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/MemberList'
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...